Status Overlay fix + refactoring

+++ common/java/com/zergatul/cheatutils/modules/esp/EntityTitle.java
@@ -210,7 +210,7 @@ public class EntityTitle implements FontBackendHolder {
             return;
         }
 
-        int scale = (int) mc.getWindow().getGuiScale(); // currently it is always integer
+        int scale = mc.getWindow().getGuiScale();
         int scrWidth = mc.getWindow().getWidth();
         int scrHeight = mc.getWindow().getHeight();
         int halfScrWidth = scrWidth / 2;
